+++ controllers/util/managerSet.go
@@ -65,9 +65,10 @@ func NewManagerSet(restConfig *rest.Config, client client.Client, clientSet kube
 
 	requirementsGenerator := resource.NewRequirementsGenerator(configRepos.DoguConfigRepository)
 	hostAliasGenerator := alias.NewHostAliasGenerator(configRepos.GlobalConfigRepository)
-	doguResourceGenerator := resource.NewResourceGenerator(client.Scheme(), requirementsGenerator, hostAliasGenerator, additionalImages)
+	securityContextGenerator := resource.NewSecurityContextGenerator()
+	doguResourceGenerator := resource.NewResourceGenerator(client.Scheme(), requirementsGenerator, hostAliasGenerator, securityContextGenerator, additionalImages)
 
-	upserter := resource.NewUpserter(client, doguResourceGenerator)
+	upserter := resource.NewUpserter(client, doguResourceGenerator, config.NetworkPoliciesEnabled)
 
 	doguRegistrator := cesregistry.NewCESDoguRegistrator(doguVersionReg, doguDescriptorRepo)
 	imageRegistry := imageregistry.NewCraneContainerImageRegistry()